requirements.txt

pandas 
streamlit
plotly
scikit-learn
